nexodus: rename internal/apex to internal/nexodus
