+++ internal/nexodus/utils_linux.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package apex
+package nexodus
 
 import (
 	"fmt"
