+++ internal/nexodus/utils_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package apex
+package nexodus
 
 import (
 	"fmt"
