+++ internal/nexodus/wg-deploy.go
@@ -1,4 +1,4 @@
-package apex
+package nexodus
 
 import (
 	"github.com/google/uuid"
