+++ internal/nexodus/wg-iface.go
@@ -1,4 +1,4 @@
-package apex
+package nexodus
 
 import (
 	"errors"
